Allow Cluster subclasses to delegate to another instance (JAVA-619).
This moves all significant work to Cluster.Manager.init(). A delegating
subclass must ensure that init() is never called in the parent class, so
that extra ressources like thread pools are not created unnecessarily.

This is a workaround for the fact that Cluster should be an interface,
which would be a better way to address this use case, but can't be done
for compatibility reasons.